+++ binder/mapping.go
@@ -34,7 +34,7 @@ var (
 	// decoderPoolMap helps to improve binders
 	decoderPoolMap = map[string]*sync.Pool{}
 	// tags is used to classify parser's pool
-	tags = []string{"header", "respHeader", "cookie", "query", "form", uriTagName, "params"}
+	tags = []string{"header", "respHeader", "cookie", "query", "form", "uri"}
 )
 
 func getDecoderPool(tag string) *sync.Pool {
@@ -261,7 +261,6 @@ var (
 	queryFieldCache      sync.Map
 	formFieldCache       sync.Map
 	uriFieldCache        sync.Map
-	paramsFieldCache     sync.Map
 )
 
 func getFieldCache(aliasTag string) *sync.Map {
@@ -274,10 +273,8 @@ func getFieldCache(aliasTag string) *sync.Map {
 		return &cookieFieldCache
 	case "form":
 		return &formFieldCache
-	case uriTagName:
+	case "uri":
 		return &uriFieldCache
-	case "params":
-		return &paramsFieldCache
 	case "query":
 		return &queryFieldCache
 	}